+++ include/mysql_connection.h
@@ -88,7 +88,7 @@ class MySQL_Connection {
 		char *ldap_user_variable_value;
 		bool ldap_user_variable_sent;
 		uint8_t protocol_version;
-		uint8_t charset;
+		unsigned int charset;
 		uint8_t sql_log_bin;
 		int8_t last_set_autocommit;
 		bool autocommit;
@@ -149,7 +149,7 @@ class MySQL_Connection {
 	~MySQL_Connection();
 	bool set_autocommit(bool);
 	bool set_no_backslash_escapes(bool);
-	uint8_t set_charset(uint8_t);
+	unsigned int set_charset(unsigned int);
 
 	void set_status_transaction(bool);
 	void set_status_compression(bool);
@@ -197,7 +197,7 @@ class MySQL_Connection {
 	int async_change_user(short event);
 	int async_select_db(short event);
 	int async_set_autocommit(short event, bool);
-	int async_set_names(short event, uint8_t nr);
+	int async_set_names(short event, unsigned int nr);
 	int async_send_simple_command(short event, char *stmt, unsigned long length); // no result set expected
 	int async_query(short event, char *stmt, unsigned long length, MYSQL_STMT **_stmt=NULL, stmt_execute_metadata_t *_stmt_meta=NULL);
 	int async_ping(short event);
